+++ node/src/Commands.ts
@@ -569,6 +569,14 @@ export function createSismember(
 /**
  * @internal
  */
+export function createSPop(key: string, count?: number): redis_request.Command {
+    const args: string[] = count == undefined ? [key] : [key, count.toString()];
+    return createCommand(RequestType.Spop, args);
+}
+
+/**
+ * @internal
+ */
 export function createCustomCommand(args: string[]) {
     return createCommand(RequestType.CustomCommand, args);
 }