+++ node/src/Transaction.ts
@@ -62,6 +62,7 @@ import {
     createSAdd,
     createSCard,
     createSMembers,
+    createSPop,
     createSRem,
     createSelect,
     createSet,
@@ -675,6 +676,32 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      */
     public sismember(key: string, member: string): T {
         return this.addAndReturn(createSismember(key, member));
+    }
+
+    /** Removes and returns one random member from the set value store at `key`.
+     * See https://redis.io/commands/spop/ for details.
+     * To pop multiple members, see `spopCount`.
+     *
+     * @param key - The key of the set.
+     *
+     * Command Response - the value of the popped member.
+     * If `key` does not exist, null will be returned.
+     */
+    public spop(key: string): T {
+        return this.addAndReturn(createSPop(key));
+    }
+
+    /** Removes and returns up to `count` random members from the set value store at `key`, depending on the set's length.
+     * See https://redis.io/commands/spop/ for details.
+     *
+     * @param key - The key of the set.
+     * @param count - The count of the elements to pop from the set.
+     *
+     * Command Response - A list of popped elements will be returned depending on the set's length.
+     * If `key` does not exist, empty list will be returned.
+     */
+    public spopCount(key: string, count: number): T {
+        return this.addAndReturn(createSPop(key, count));
     }
 
     /** Returns the number of keys in `keys` that exist in the database.